Use convert_from_extras before ignore_missing
In the *_show schema the convert_from_extras converter should be placed
first, as at this point the field does not yet exist and with
ignore_missing the whole process is cancelled. Only if the order is like
that, we actually get a new field image_display_url as intended.

Signed-off-by: blagoja <blagojabozinovski@gmail.com>